pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["clockify_core", "api_studio", "universal_webhook"]

[project]
name = "clockify-api-studio"
version = "0.1.0"
description = "Python backend for Clockify API Studio add-on"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "SQLAlchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "tenacity>=8.2.0",
    "jsonpath-ng>=1.6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]
